feat(pci-kubernetes): move domain mappers from viewmodel to service
ref: #TAPC-5549
